separate Sprites class
